[controller] fix input delay when there are a lot data to read.
the problem is that the controller will save all items read before send the
NewItem event to matcher. Meanwhile the matcher keeps going because no signal
is received. And since the controller will need to lock the data of items and
that lock is not released by matcher. Kind of deadlock here.