[model] save selected items' indics in model.
Previously whether an item is selected or not is saved in the item. Normally
the selected item won't be much. When finally we want to output, the original
strategy would have to iterate through all existing items and check whether
they are selected or not.

Now we extract this info out, so that we only need to iterate over the
selected ones instead of the whole set. Improve the speed.