[orderedvec] give up on binaryheap
Binary Heap will increase the overhead on insertion(not much thought) and it
did not guarentee the output order, thus it will cause the screen flickr when
new items are read and matched.

Now we use a small ordered vector + a large unordered vector and sort the
unordered vector on demand.