[speed] share matched_items between matcher and model
Before: we will cache the result of a search for later use, this will cause us
to clone the result even it is the first time to search. If the matched item
number is large(short query) then the clone time is large(30%).

After: we share the matched result so this clone time should be reduced.

Rust did not support sharing reference between threads to keep thread safty.
Thus we use Arc<RwLock<_>> to wrap up the data and send the Arc throught
eventbox. The Cons of this method I think is making fields dirty(look ugly).