[enhance] reduce screen flickr for interactive mode
In normal mode, the items are fetched in one run and matched for different
queries. In interactive mode, the items are feched every time, and usually the
rate of the input cannot be predicted(for example, calling `ag` in a large
project). Thus will more easily cause screen flicker.

Also more synchrinization between threads should be made than normal mode.