fix #70 allow specify multiple options
Previously the options that takes values had been added supported. At
that time I did not add support for the rest because some of them are
actually conflict, for example `--multiple` and `--no-multi`. And
clap.rs did not support finding the last occurred options in group.

Then I find that ony two group needs special care:
- multiple vs no-multiple
- height vs no-height

So I decide that if the `no-` options is occurred, then the feature is
just disabled. That's all, adding `multiple` or `height` cannot cancel
the effect.