Fix references to SKIM_DEFAULT_OPTS in vim plugin
There are references to SKIM_DEFAULT_OPTS within the vim plugin, but this is different from the SKIM_DEFAULT_OPTIONS that is suggested for use within the documentation, and used elsewhere throughout the program.  This changes the two uses of SKIM_DEFAULT_OPTS to SKIM_DEFAULT_OPTIONS within the vim plugin.