fix #123 100% CPU when input is not ready
It happens because sender is checking infinitely for the new inputs. Now
we set some timeout to fix this issue.

However the example given by the issue `cat | sk` still won't work
because skim has no way to kill the reader thread directly(limitation of
thread APIs). Also the input won't generate EOF because `cat` is waiting
for user input.

I decide not to solve this issue for now, basically because I think it
rarely happens in real use cases.