[orderedvec] defer merging when fetching
Previous when inserting, orderedvec would perform a global sorting which
might block the main thread with large number of items.  Also
unnecessary sorting are performed when new items are added.

This commit would only sort the vector that is append, and merging the
partially sorted items only when fetching. Normally user would only
access the first few items(~= 100), so that the global sorting cost is
reduced in most cases.

Note that this version would use single thread for sorting vectors, the
performance may or may not be reduced compared to the previous
multi-thread version.