fix #344: implement text() and display() correctly
The finally "display" of an item was drawn by first invoking `display()`
for the content, and then draw the matched indices. When the customized
`display()` did not return the same text as `text()`, the display is
broken.

PR #347 try to fix this by adding a "non-highlight" option. But maybe
it's better to hand over all the drawing tasks to `display()`, including
highlight the matching indices.