fix #365: new option `--show-cmd-error`
Currently the invoked command would fail silently. Sometimes this might
not be desirable because we have no idea whether it outputs nothing or
died.

New option `--show-cmd-error` is designed to send the error message and
display them as items.

It's somehow a trade-off choice that I don't want to modify the
current library interface. Also it's not made default because there
might be several fail invocation before we complete the whole
"cmd-query".